+++ CHANGELOG.md
@@ -18,6 +18,10 @@ Security in case of vulnerabilities.
  - New route `/projects/team/completed`
  - Your team projects completed list (`/projects/team/completed`)
 
+### Changed
+ - Filter out any local authorities with no projects in `ListAllProjectByLocalAuthorities`
+
+ 
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/main...production-2025-04-17.164) for everything awaiting release
 
 ---